# Add generic dimension handling interface