--- a/processing/src/main/java/io/druid/query/dimension/BaseFilteredDimensionSpec.java
+++ b/processing/src/main/java/io/druid/query/dimension/BaseFilteredDimensionSpec.java
@@ -23,6 +23,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.common.base.Preconditions;
 import io.druid.query.extraction.ExtractionFn;
 import io.druid.segment.DimensionSelector;
+import io.druid.segment.column.ColumnCapabilities;
 import io.druid.segment.data.IndexedInts;
 import io.druid.segment.data.ListBasedIndexedInts;
 
@@ -116,6 +117,12 @@ public abstract class BaseFilteredDimensionSpec implements DimensionSpec
       public int lookupId(String name)
       {
         return forwardMapping.get(selector.lookupId(name));
+      }
+
+      @Override
+      public ColumnCapabilities getDimCapabilities()
+      {
+        return selector.getDimCapabilities();
       }
     };
   }