--- a/processing/src/main/java/io/druid/query/groupby/GroupByQueryEngine.java
+++ b/processing/src/main/java/io/druid/query/groupby/GroupByQueryEngine.java
@@ -40,16 +40,18 @@ import io.druid.collections.StupidPool;
 import io.druid.data.input.MapBasedRow;
 import io.druid.data.input.Row;
 import io.druid.guice.annotations.Global;
-import io.druid.query.Query;
 import io.druid.query.aggregation.AggregatorFactory;
 import io.druid.query.aggregation.BufferAggregator;
 import io.druid.query.aggregation.PostAggregator;
 import io.druid.query.dimension.DimensionSpec;
 import io.druid.query.filter.Filter;
 import io.druid.segment.Cursor;
+import io.druid.segment.DimensionHandler;
 import io.druid.segment.DimensionSelector;
 import io.druid.segment.StorageAdapter;
-import io.druid.segment.data.IndexedInts;
+import io.druid.segment.StringDimensionHandler;
+import io.druid.segment.column.Column;
+import io.druid.segment.column.ColumnCapabilities;
 import io.druid.segment.filter.Filters;
 import org.joda.time.DateTime;
 import org.joda.time.Interval;
@@ -58,7 +60,7 @@ import javax.annotation.Nullable;
 import java.io.Closeable;
 import java.io.IOException;
 import java.nio.ByteBuffer;
-import java.util.ArrayList;
+import java.util.Comparator;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
@@ -72,6 +74,66 @@ public class GroupByQueryEngine
 
   private final Supplier<GroupByQueryConfig> config;
   private final StupidPool<ByteBuffer> intermediateResultsBufferPool;
+
+  private static final DimensionHandler TIME_AND_NULL_DIMENSION_HANDLER = new StringDimensionHandler("time-null");
+
+  /*
+   * Relative reads change the current position of a ByteBuffer.
+   * The key comparator uses absolute reads to avoid changing the state of the ByteBuffer.
+   * ByteBuffer does not provide an absolute bulk get() method, so this method provides that functionality.
+   */
+  private static void getBytesFromBuffer(ByteBuffer src, byte[] dst, int srcOffset, int readLen)
+  {
+    for (int i = 0; i < readLen; i++) {
+      dst[i] = src.get(srcOffset + i);
+    }
+  }
+
+  private static final Comparator<ByteBuffer> makeKeyComparator(final List<GroupByDimensionInfo> dimInfo)
+  {
+    Comparator comp = new Comparator<ByteBuffer>()
+    {
+      public int compare(ByteBuffer o1, ByteBuffer o2)
+      {
+        int pos = 0;
+        int limit = o1.limit();
+        int ret = 0;
+        int dimIndex = 0;
+        int maxDimIndex = dimInfo.size();
+
+        while (pos < limit && dimIndex < maxDimIndex) {
+          DimensionHandler handler = dimInfo.get(dimIndex).handler;
+          if (handler == null) {
+            // Special time column handling
+            int int1 = o1.getInt(pos);
+            int int2 = o2.getInt(pos);
+            pos += Ints.BYTES;
+            ret = Integer.compare(int1, int2);
+          } else {
+            int valLen = handler.getEncodedValueSize();
+            byte[] bytes1 = new byte[valLen];
+            byte[] bytes2 = new byte[valLen];
+
+            getBytesFromBuffer(o1, bytes1, pos, valLen);
+            getBytesFromBuffer(o2, bytes2, pos, valLen);
+
+            pos += valLen;
+
+            ret = handler.compareEncodedValueAsBytes(bytes1, bytes2);
+          }
+
+          if (ret != 0) {
+            return ret;
+          }
+
+          dimIndex++;
+        }
+        return ret;
+      }
+
+    };
+    return comp;
+  }
 
   @Inject
   public GroupByQueryEngine(
@@ -122,7 +184,7 @@ public class GroupByQueryEngine
                           @Override
                           public RowIterator make()
                           {
-                            return new RowIterator(query, cursor, bufferHolder.get(), config.get());
+                            return new RowIterator(query, cursor, bufferHolder.get(), config.get(), storageAdapter);
                           }
 
                           @Override
@@ -152,20 +214,25 @@ public class GroupByQueryEngine
     private final ByteBuffer metricValues;
     private final BufferAggregator[] aggregators;
     private final PositionMaintainer positionMaintainer;
+    private final StorageAdapter adapter;
 
-    private final Map<ByteBuffer, Integer> positions = Maps.newTreeMap();
+    private final Map<ByteBuffer, Integer> positions;
     // GroupBy queries tend to do a lot of reads from this. We co-store a hash map to make those reads go faster.
     private final Map<ByteBuffer, Integer> positionsHash = Maps.newHashMap();
 
     public RowUpdater(
         ByteBuffer metricValues,
         BufferAggregator[] aggregators,
-        PositionMaintainer positionMaintainer
+        PositionMaintainer positionMaintainer,
+        StorageAdapter adapter,
+        List<GroupByDimensionInfo> dimInfo
     )
     {
       this.metricValues = metricValues;
       this.aggregators = aggregators;
       this.positionMaintainer = positionMaintainer;
+      this.adapter = adapter;
+      this.positions = Maps.newTreeMap(makeKeyComparator(dimInfo));
     }
 
     public int getNumRows()
@@ -180,26 +247,29 @@ public class GroupByQueryEngine
 
     private List<ByteBuffer> updateValues(
         ByteBuffer key,
-        List<DimensionSelector> dims
+        final List<GroupByDimensionInfo> dimInfo
     )
     {
-      if (dims.size() > 0) {
+      if (dimInfo.size() > 0) {
         List<ByteBuffer> retVal = null;
         List<ByteBuffer> unaggregatedBuffers = null;
+        GroupByDimensionInfo info = dimInfo.get(0);
 
-        final DimensionSelector dimSelector = dims.get(0);
-        final IndexedInts row = dimSelector.getRow();
-        if (row == null || row.size() == 0) {
-          ByteBuffer newKey = key.duplicate();
-          newKey.putInt(dimSelector.getValueCardinality());
-          unaggregatedBuffers = updateValues(newKey, dims.subList(1, dims.size()));
-        } else {
-          for (Integer dimValue : row) {
-            ByteBuffer newKey = key.duplicate();
-            newKey.putInt(dimValue);
-            unaggregatedBuffers = updateValues(newKey, dims.subList(1, dims.size()));
+        final DimensionSelector dimSelector = info.selector;
+        DimensionHandler dimHandler = info.handler;
+
+        final List<GroupByDimensionInfo> dimInfoSublist = dimInfo.subList(1, dimInfo.size());
+        final Function<ByteBuffer, List<ByteBuffer>> updateValuesFn = new Function<ByteBuffer, List<ByteBuffer>>()
+        {
+          @Override
+          public List<ByteBuffer> apply(ByteBuffer input)
+          {
+            return updateValues(input, dimInfoSublist);
           }
-        }
+        };
+
+        unaggregatedBuffers = dimHandler.groupByUpdater(dimSelector, key, updateValuesFn);
+
         if (unaggregatedBuffers != null) {
           if (retVal == null) {
             retVal = Lists.newArrayList();
@@ -290,6 +360,30 @@ public class GroupByQueryEngine
     }
   }
 
+  private static class GroupByDimensionInfo
+  {
+    final DimensionSelector selector;
+    final String outputName;
+    final String originalName;
+    final ColumnCapabilities capabilities;
+    final DimensionHandler handler;
+
+    public GroupByDimensionInfo(
+        DimensionSelector selector,
+        String outputName,
+        String originalName,
+        ColumnCapabilities capabilities,
+        DimensionHandler handler
+    )
+    {
+      this.selector = selector;
+      this.outputName = outputName;
+      this.originalName = originalName;
+      this.capabilities = capabilities;
+      this.handler = handler;
+    }
+  }
+
   private static class RowIterator implements CloseableIterator<Row>
   {
     private final GroupByQuery query;
@@ -298,21 +392,23 @@ public class GroupByQueryEngine
     private final int maxIntermediateRows;
 
     private final List<DimensionSpec> dimensionSpecs;
-    private final List<DimensionSelector> dimensions;
-    private final ArrayList<String> dimNames;
     private final List<AggregatorFactory> aggregatorSpecs;
     private final BufferAggregator[] aggregators;
     private final String[] metricNames;
     private final int[] sizesRequired;
+    private final StorageAdapter adapter;
+    private final List<GroupByDimensionInfo> dimInfo;
 
     private List<ByteBuffer> unprocessedKeys;
     private Iterator<Row> delegate;
+    private int keyBufferSize;
 
-    public RowIterator(GroupByQuery query, final Cursor cursor, ByteBuffer metricsBuffer, GroupByQueryConfig config)
+    public RowIterator(GroupByQuery query, final Cursor cursor, ByteBuffer metricsBuffer, GroupByQueryConfig config, StorageAdapter adapter)
     {
       this.query = query;
       this.cursor = cursor;
       this.metricsBuffer = metricsBuffer;
+      this.adapter = adapter;
 
       this.maxIntermediateRows = Math.min(
           query.getContextValue(
@@ -324,15 +420,34 @@ public class GroupByQueryEngine
       unprocessedKeys = null;
       delegate = Iterators.emptyIterator();
       dimensionSpecs = query.getDimensions();
-      dimensions = Lists.newArrayListWithExpectedSize(dimensionSpecs.size());
-      dimNames = Lists.newArrayListWithExpectedSize(dimensionSpecs.size());
+      dimInfo = Lists.newArrayListWithExpectedSize(dimensionSpecs.size());
+
+      Map<String, DimensionHandler> handlerMap = adapter.getDimensionHandlers();
 
       for (int i = 0; i < dimensionSpecs.size(); ++i) {
         final DimensionSpec dimSpec = dimensionSpecs.get(i);
         final DimensionSelector selector = cursor.makeDimensionSelector(dimSpec);
         if (selector != null) {
-          dimensions.add(selector);
-          dimNames.add(dimSpec.getOutputName());
+          String dimName = dimSpec.getDimension();
+          DimensionHandler handler = handlerMap.get(dimName);
+          if (handler == null) {
+            handler = TIME_AND_NULL_DIMENSION_HANDLER;
+          }
+          GroupByDimensionInfo info = new GroupByDimensionInfo(
+              selector,
+              dimSpec.getOutputName(),
+              dimName,
+              adapter.getColumnCapabilities(dimName),
+              handler
+          );
+
+          dimInfo.add(info);
+
+          if (dimName.equals(Column.TIME_COLUMN_NAME) || handler == null) {
+            keyBufferSize += Ints.BYTES;
+          } else {
+            keyBufferSize += handler.getEncodedValueSize();
+          }
         }
       }
 
@@ -346,6 +461,7 @@ public class GroupByQueryEngine
         metricNames[i] = aggregatorSpec.getName();
         sizesRequired[i] = aggregatorSpec.getMaxIntermediateSize();
       }
+
     }
 
     @Override
@@ -366,10 +482,10 @@ public class GroupByQueryEngine
       }
 
       final PositionMaintainer positionMaintainer = new PositionMaintainer(0, sizesRequired, metricsBuffer.remaining());
-      final RowUpdater rowUpdater = new RowUpdater(metricsBuffer, aggregators, positionMaintainer);
+      final RowUpdater rowUpdater = new RowUpdater(metricsBuffer, aggregators, positionMaintainer, adapter, dimInfo);
       if (unprocessedKeys != null) {
         for (ByteBuffer key : unprocessedKeys) {
-          final List<ByteBuffer> unprocUnproc = rowUpdater.updateValues(key, ImmutableList.<DimensionSelector>of());
+          final List<ByteBuffer> unprocUnproc = rowUpdater.updateValues(key, ImmutableList.<GroupByDimensionInfo>of());
           if (unprocUnproc != null) {
             throw new ISE("Not enough memory to process the request.");
           }
@@ -377,13 +493,11 @@ public class GroupByQueryEngine
         cursor.advance();
       }
       while (!cursor.isDone() && rowUpdater.getNumRows() < maxIntermediateRows) {
-        ByteBuffer key = ByteBuffer.allocate(dimensions.size() * Ints.BYTES);
-
-        unprocessedKeys = rowUpdater.updateValues(key, dimensions);
+        ByteBuffer key = ByteBuffer.allocate(keyBufferSize);
+        unprocessedKeys = rowUpdater.updateValues(key, dimInfo);
         if (unprocessedKeys != null) {
           break;
         }
-
         cursor.advance();
       }
 
@@ -408,12 +522,10 @@ public class GroupByQueryEngine
                   Map<String, Object> theEvent = Maps.newLinkedHashMap();
 
                   ByteBuffer keyBuffer = input.getKey().duplicate();
-                  for (int i = 0; i < dimensions.size(); ++i) {
-                    final DimensionSelector dimSelector = dimensions.get(i);
-                    final int dimVal = keyBuffer.getInt();
-                    if (dimSelector.getValueCardinality() != dimVal) {
-                      theEvent.put(dimNames.get(i), dimSelector.lookupName(dimVal));
-                    }
+                  for (int i = 0; i < dimInfo.size(); ++i) {
+                    final GroupByDimensionInfo info = dimInfo.get(i);
+                    final DimensionHandler dimHandler = info.handler;
+                    dimHandler.addValueToEventFromGroupByKey(keyBuffer, info.selector, theEvent, info.outputName);
                   }
 
                   int position = input.getValue();