--- a/processing/src/main/java/io/druid/query/topn/AlphaNumericTopNMetricSpec.java
+++ b/processing/src/main/java/io/druid/query/topn/AlphaNumericTopNMetricSpec.java
@@ -34,7 +34,7 @@ public class AlphaNumericTopNMetricSpec extends LexicographicTopNMetricSpec
 {
   private static final byte CACHE_TYPE_ID = 0x2;
 
-  protected static Comparator<String> comparator = StringComparators.ALPHANUMERIC;
+  protected static Comparator<Comparable> comparator = StringComparators.ALPHANUMERIC;
 
   @JsonCreator
   public AlphaNumericTopNMetricSpec(