@jon-wei 2016-03-31
rebasing, there are some conflicts with https://github.com/druid-io/druid/pull/2753