@jon-wei 2016-03-31
rebased

@gianm Updated StringDimensionHandler with your recent IndexMerger changes; it had most of the null handling code already since StringDimensionMerger was based off IndexMergerV9, just had to add this part to closeWritersV8():

https://github.com/jon-wei/druid/blob/dimension_interface/processing/src/main/java/io/druid/segment/StringDimensionHandler.java#L752