@jon-wei 2016-04-25
Added some benchmark numbers comparing performance between this patch and master, as of 4/25/2016