@xvrl 2016-04-26
@jon-wei based on your numbers it looks like most of the queries are unaffected, however I do see a 20% increase in `SearchBenchmark.queryIncrementalIndex` and between 10% and 15% increase in `GroupByBenchmark.queryIncrementalIndex` and `GroupByBenchmark.queryQueryableIndex`.

It would be good to investigate those in more detail and see if we can reduce the impact on those queries. 10% is a pretty significant drop in performance. 

Everything else is within +/- 5%, which is with error bands. You might want to run your benchmark on EC2 instead of a laptop to get those errors bands narrower (see https://metamarkets.com/2015/effect-of-frequency-governor-on-java-benchmarking/)