@jon-wei 2016-04-28
@xvrl I've been digging into the performance slowdown in GroupBy and Search.

I've made some improvements to GroupByEngine (haven't committed changes to this PR yet).

The increase shown in SearchBenchmark was found to ultimately come from an increase in the time it takes to read rows from an IncrementalIndex. 

I've narrowed the cause there to the change from int[][] to Comparable[][] in various places, but still looking into how to improve the performance there.