@jon-wei 2016-05-10
Updated this to address PR comments, will post benchmark results when I have them available.