@jon-wei 2016-05-12
Updated description with new benchmark results as of May 12. 

The description also mentions a change in the locking behavior of the dimension value dictionary, with related benchmark results.