@jon-wei 2016-07-01
Going to break this PR up even further, closing this for now